Version 2.4.0 Dependencies upgrade: Kotlin 1.3.60, serialization 0.14.0, ktor 1.2.6, spring boot 2.2.1 and others

+++ kvision-modules/kvision-bootstrap-upload/src/main/kotlin/pl/treksoft/kvision/form/upload/UploadInput.kt
@@ -268,7 +268,7 @@ open class UploadInput(uploadUrl: String? = null, multiple: Boolean = false, cla
         val list = mutableListOf<KFile>()
         for (key in js("Object").keys(fileStack)) {
             val nativeFile = fileStack[key].file as File
-            val kFile = KFile(nativeFile.name, nativeFile.size, null)
+            val kFile = KFile(nativeFile.name, nativeFile.size.toInt(), null)
             nativeFiles[kFile] = nativeFile
             list += kFile
         }
